// apps/worker/src/core/types.ts
export interface User {
  user_id: string;
  username: string;
  wechat_name: string | null;
  role: 'member' | 'moderator' | 'admin';
  power: number;
  is_active: 0 | 1;
  last_username_change_at_utc: string | null;
  deleted_at_utc: string | null;
  created_at_utc: string;
  updated_at_utc: string;
}